+++ src/main/java/chess/controller/SpringExceptionAdvice.java
@@ -1,0 +1,17 @@
+package chess.controller;
+
+import java.util.Map;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@ControllerAdvice
+public class SpringExceptionAdvice {
+    @ExceptionHandler(RuntimeException.class)
+    public ResponseEntity<Map<String, String>> handle(RuntimeException exception) {
+        return ResponseEntity.badRequest().body(Map.of(
+            "exception", exception.getMessage()
+        ));
+    }
+}
