+++ src/main/java/chess/controller/SpringRestChessController.java
@@ -4,14 +4,11 @@ import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
 
 import chess.domain.game.GameState;
 import chess.dto.Arguments;
@@ -20,51 +17,34 @@ import chess.dto.GameStateResponse;
 import chess.dto.PathResponse;
 import chess.service.GameService;
 
-@Controller
-public class SpringChessController {
+@RestController
+public class SpringRestChessController {
 
     private final GameService gameService;
 
     @Autowired
-    public SpringChessController(GameService gameService) {
+    public SpringRestChessController(GameService gameService) {
         this.gameService = gameService;
     }
 
-    @GetMapping("/")
-    public String showIndex() {
-        return "index";
-    }
-
-    @GetMapping(path = "/main")
-    public String showGame(Model model, @RequestParam("room_name") String roomName) {
-        GameState state = gameService.readGameState(roomName);
-        GameStateResponse response = toGameStateResponse(state);
-        model.addAttribute("response", response);
-        return "game";
-    }
-
     @GetMapping(path = "/create")
-    @ResponseBody
     public ResponseEntity<PathResponse> create(@RequestParam("room_name") String roomName) {
         gameService.createNewGame(roomName);
         return respondPath("/main?room_name=" + roomName);
     }
 
     @GetMapping(path = "/enter")
-    @ResponseBody
     public ResponseEntity<PathResponse> enter(@RequestParam("room_name") String roomName) {
         return respondPath("/main?room_name=" + roomName);
     }
 
     @GetMapping(path = "/start")
-    @ResponseBody
     public ResponseEntity<PathResponse> start(@RequestParam("room_name") String roomName) {
         gameService.startGame(roomName);
         return respondPath("/main?room_name=" + roomName);
     }
 
     @GetMapping(path = "/end")
-    @ResponseBody
     public ResponseEntity<PathResponse> end(@RequestParam("room_name") String roomName) {
         gameService.finishGame(roomName);
         return respondPath("/");
@@ -75,7 +55,7 @@ public class SpringChessController {
         @RequestBody String body) {
         final Arguments arguments = Arguments.ofJson(body, Command.MOVE.getParameters());
         GameState state = gameService.moveBoard(roomName, arguments);
-        return ResponseEntity.ok().body(toGameStateResponse(state));
+        return ResponseEntity.ok().body(GameStateResponse.of(state));
     }
 
     @GetMapping(path = "/status")
@@ -86,18 +66,7 @@ public class SpringChessController {
         );
     }
 
-    @ExceptionHandler(RuntimeException.class)
-    public ResponseEntity<Map<String, String>> handle(RuntimeException exception) {
-        return ResponseEntity.badRequest().body(Map.of(
-            "exception", exception.getMessage()
-        ));
-    }
-
     private ResponseEntity<PathResponse> respondPath(String path) {
         return ResponseEntity.ok().body(new PathResponse(path));
-    }
-
-    private GameStateResponse toGameStateResponse(GameState state) {
-        return new GameStateResponse(BoardResponse.of(state.getPointPieces()), state.getState(), state.getColor());
     }
 }
