+++ src/main/java/chess/dto/GameStateResponse.java
@@ -1,5 +1,7 @@
 package chess.dto;
 
+import chess.domain.game.GameState;
+
 public class GameStateResponse {
     private final BoardResponse board;
     private final String state;
@@ -9,6 +11,10 @@ public class GameStateResponse {
         this.board = board;
         this.state = state;
         this.color = color;
+    }
+
+    public static GameStateResponse of(GameState state) {
+        return new GameStateResponse(BoardResponse.of(state.getPointPieces()), state.getState(), state.getColor());
     }
 
     public BoardResponse getBoard() {
